+++ src/Rect.cpp
@@ -7,7 +7,7 @@ Rect::Rect() {
   this->height = 0;
 }
 
-Rect::Rect(int x, int y, int width, int height) {
+Rect::Rect(float x, float y, float width, float height) {
   this->x = x;
   this->y = y;
   this->width = width;
@@ -28,10 +28,10 @@ bool Rect::WithinYLimits(float yPosition) {
 }
 
 Vec2 Rect::GetCenter() {
-  return(Vec2(this->x + (this->width / 2), this->y + (this->height / 2)));
+  return(Vec2(this->x + (this->width / 2.0), this->y + (this->height / 2.0)));
 }
 
 void Rect::SetCenter(Vec2 center) {
-  this->x = center.x - (this->width / 2);
-  this->y = center.y - (this->height / 2);
+  this->x = center.x - (this->width / 2.0);
+  this->y = center.y - (this->height / 2.0);
 }
