+++ src/TileMap.cpp
@@ -37,7 +37,7 @@ void TileMap::SetTileSet(TileSet* tileSet) {
   this->tileSet = tileSet;
 }
 
-void TileMap::RenderLayer(int layer, int cameraX, int cameraY) {
+void TileMap::RenderLayer(int layer, float cameraX, float cameraY) {
   for (int j = 0 ; j < this->mapHeight ; j++) {
     for (int i = 0 ; i < this->mapWidth ; i++) {
       this->tileSet->RenderTile(
